Correctly deserialize empty strings into empty sequence
This change fixes the deserialization of a sequence (most likely
transformed into a Vec later) when an empty string is provided. So far
the result has been a sequence with a single empty string (or an error
when directly attempting to convert it further, say, into an integer).
It should be uncontroversial that an empty string (i.e., one without any
separator) represents an empty sequence and not a sequence with an
"empty" element.
To that end, this change adjusts the deserialization code to fix this
case.

Fixes: #50